+++ src/Dignite.Site.Admin.HttpApi.Client/SiteAdminHttpApiClientModule.cs
@@ -14,7 +14,7 @@ public class SiteAdminHttpApiClientModule : AbpModule
 {
     public override void ConfigureServices(ServiceConfigurationContext context)
     {
-        context.Services.AddHttpClientProxies(
+        context.Services.AddStaticHttpClientProxies(
             typeof(SiteAdminApplicationContractsModule).Assembly,
             SiteAdminRemoteServiceConsts.RemoteServiceName
         );