+++ src/Dignite.Site.Public.HttpApi.Client/SitePublicHttpApiClientModule.cs
@@ -14,7 +14,7 @@ public class SitePublicHttpApiClientModule : AbpModule
 {
     public override void ConfigureServices(ServiceConfigurationContext context)
     {
-        context.Services.AddHttpClientProxies(
+        context.Services.AddStaticHttpClientProxies(
             typeof(SitePublicApplicationContractsModule).Assembly,
             SitePublicRemoteServiceConsts.RemoteServiceName
         );